requirements.txt
django>=4.2
psycopg2-binary==2.9.9
redis==5.0.1
pillow
django-debug-toolbar
stripe
Pillow
celery
channels
daphne
channels-redis
pytest-django
pytest-cov
factory-boy